google-datastore-gds/requirements.txt
google-cloud-datastore==2.6.0
